+++ R/send_password_reset_email_module.R
@@ -7,6 +7,8 @@
 #' @importFrom shiny actionLink NS
 #' @importFrom shinyFeedback useShinyFeedback
 #'
+#' @return the UI to create a password reset link.
+#'
 #' @export
 send_password_reset_email_module_ui <- function(id, link_text = "Forgot your password?") {
   ns <- shiny::NS(id)
@@ -35,6 +37,8 @@ send_password_reset_email_module_ui <- function(id, link_text = "Forgot your pas
 #' @importFrom httr POST authenticate content status_code
 #' @importFrom jsonlite fromJSON
 #' @importFrom shinyFeedback showToast
+#'
+#' @return \code{invisible(NULL)}
 #'
 #' @export
 #'
@@ -79,15 +83,17 @@ send_password_reset_email_module <- function(input, output, session, email) {
         .options = polished_toast_options
       )
     }, error = function(err) {
-      print(err)
+      err_msg <- conditionMessage(err)
+      warning(err_msg)
       shinyFeedback::showToast(
         "error",
-        err$message,
+        err_msg,
         .options = polished_toast_options
       )
     })
 
   })
 
+  invisible(NULL)
 }